// api/generated/model/createMyRecipeBody.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * A Cheffy API é o backend central de uma aplicação de receitas, projetado para gerenciar catálogo culinário, categorias, tags, ingredientes, informações nutricionais, favoritos de usuário, autenticação via Better Auth, assinatura de upload no Cloudinary e um assistente gastronômico integrado à Groq.
 * OpenAPI spec version: 1.0.0
 */
import type { CreateMyRecipeBodyDifficulty } from './createMyRecipeBodyDifficulty';
import type { CreateMyRecipeBodySectionsItem } from './createMyRecipeBodySectionsItem';
import type { CreateMyRecipeBodyYieldUnit } from './createMyRecipeBodyYieldUnit';

export type CreateMyRecipeBody = {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imagePublicId?: string | null;
  /** @exclusiveMinimum 0 */
  prepTime: number;
  /** @exclusiveMinimum 0 */
  cookTime: number;
  /** @exclusiveMinimum 0 */
  yieldAmount: number;
  yieldUnit?: CreateMyRecipeBodyYieldUnit;
  difficulty?: CreateMyRecipeBodyDifficulty;
  categoryId: string;
  /** @minItems 1 */
  sections: CreateMyRecipeBodySectionsItem[];
  tagIds?: string[];
};